ZooGoN: Gulf of Naples 40-Year Zooplankton Biodiversity Assessment and Standardization

Standardizes taxonomic names in zooplankton datasets from the LTER-MareChiara station in the Gulf of Naples (1984-2024). This package is part of the DTO-BioFlow project (Digital Twin Ocean - Biodiversity Flow Integration) under the EU Horizon Mission to restore ocean and waters by 2030. It processes 40 years of Mediterranean zooplankton monitoring data, extracting standardized genus-species information and converting datasets to Darwin Core Archive format for integration with EMODnet Biology and the European Digital Twin of the Ocean.

See also

Author

Maintainer: Lorenzo Longobardi lorenzo.longobardi@gmail.com (ORCID)

Authors:

Iole Di Capua iole.dicapua@szn.it (ORCID)